Add the IRC Connection class
This is the class that does the heavy lifting as far as IRC is concerned. It
manages the connection, ensures that we don't ping out, lets us join, parses
messages from the server (currently not the full range), and lets us register a
callback function that is called in a separate thread whenever something is
said.